# docs/openapi.yaml
openapi: 3.1.0
info:
    title: RSTC API
    version: 1.0.0
    description: REST API for RSTC application.
servers:
    - url: /api
paths:
...
    /personnel:
        post:
            tags:
                - Personnel
            summary: Create personnel record
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PersonnelCreate'
                        examples:
                            create:
                                value:
                                    name: علی
                                    lname: رضایی
                                    father_name: محمد
                                    national_id: '0123456789'
                                    emp_num: '1001'
                                    hire_date: '1403-01-01'
                                    emp_type: Permanent
                                    org_post: IT
                                    job_title: برنامه‌نویس
                                    last_degree: کارشناسی
                                    phone: '09123456789'
                                    address: تهران
                                    status: فعال
                                    notes: '-'
            responses:
                '200':
                    description: Personnel created
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessResponse'
                '400':
                    $ref: '#/components/responses/Error'
                '500':
                    $ref: '#/components/responses/Error'
        get:
            tags:
                - Personnel
            summary: List personnel
            security:
                - bearerAuth: []
            responses:
                '200':
                    description: Personnel list
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Personnel'
                '500':
                    $ref: '#/components/responses/Error'
    /personnel/{id}:
        put:
            tags:
                - Personnel
            summary: Update personnel
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PersonnelUpdate'
                        examples:
                            update:
                                value:
                                    name: علی
                                    lname: رضایی
                                    father_name: محمد
                                    national_id: '0123456789'
                                    emp_num: '1001'
                                    hire_date: '1403-01-01'
                                    emp_type: Permanent
                                    org_post: IT
                                    job_title: برنامه‌نویس
                                    last_degree: کارشناسی
                                    phone: '09123456789'
                                    address: تهران
                                    status: فعال
                                    notes: '-'
            responses:
                '200':
                    description: Personnel updated
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessResponse'
                '400':
                    $ref: '#/components/responses/Error'
                '404':
                    $ref: '#/components/responses/Error'
                '500':
                    $ref: '#/components/responses/Error'
        delete:
            tags:
                - Personnel
            summary: Delete personnel
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
            responses:
                '200':
                    description: Personnel deleted
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessResponse'
                '500':
                    $ref: '#/components/responses/Error'
    /personnel/bulk:
        post:
            tags:
                - Personnel
            summary: Bulk import personnel
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: array
                            items:
                                $ref: '#/components/schemas/PersonnelCreate'
                            maxItems: 1000
                        examples:
                            bulk:
                                value:
                                    - name: علی
                                      lname: رضایی
                                      father_name: محمد
                                      national_id: '0123456789'
                                      emp_num: '1001'
                                      hire_date: '1403-01-01'
                                      emp_type: Permanent
                                      org_post: IT
                                      job_title: برنامه‌نویس
                                      last_degree: کارشناسی
                                      phone: '09123456789'
                                      address: تهران
                                      status: فعال
                                      notes: '-'
            responses:
                '200':
                    description: Import result
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                    imported:
                                        type: integer
                                    failed:
                                        type: integer
                                    errors:
                                        type: array
                                        items:
                                            type: string
                            examples:
                                result:
                                    value:
                                        success: true
                                        imported: 1
                                        failed: 0
                                        errors: []
                '400':
                    $ref: '#/components/responses/Error'
                '500':
                    $ref: '#/components/responses/Error'
...
    /options:
        get:
            tags:
                - Options
            summary: List all options
            security:
                - bearerAuth: []
            responses:
                '200':
                    description: Options map
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties:
                                    $ref: '#/components/schemas/Option'
                            examples:
                                options:
                                    value:
                                        emp_type:
                                            label: نوع استخدام
                                            options:
                                                - Permanent
                                                - Contractual
                                        mission_type:
                                            label: نوع ماموریت
                                            options:
                                                - Repair
                                                - Install
                '500':
                    $ref: '#/components/responses/Error'
    /options/{field}:
        get:
            tags:
                - Options
            summary: Get option by field
            security:
                - bearerAuth: []
            parameters:
                - name: field
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                '200':
                    description: Option detail
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Option'
                            examples:
                                option:
                                    value:
                                        label: نوع استخدام
                                        options:
                                            - Permanent
                                            - Contractual
                '404':
                    $ref: '#/components/responses/Error'
                '500':
                    $ref: '#/components/responses/Error'
        post:
            tags:
                - Options
            summary: Create option value
            security:
                - bearerAuth: []
            parameters:
                - name: field
                  in: path
                  required: true
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - value
                            properties:
                                label:
                                    type: string
                                value:
                                    type: string
                            examples:
                                create:
                                    value:
                                        label: نوع استخدام
                                        value: Permanent
            responses:
                '200':
                    description: Option created
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                    options:
                                        type: array
                                        items:
                                            type: string
                            examples:
                                created:
                                    value:
                                        success: true
                                        options:
                                            - Permanent
                '400':
                    $ref: '#/components/responses/Error'
                '500':
                    $ref: '#/components/responses/Error'
        put:
            tags:
                - Options
            summary: Update option value
            security:
                - bearerAuth: []
            parameters:
                - name: field
                  in: path
                  required: true
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - oldValue
                                - newValue
                            properties:
                                oldValue:
                                    type: string
                                newValue:
                                    type: string
                                label:
                                    type: string
                            examples:
                                update:
                                    value:
                                        oldValue: Permanent
                                        newValue: FullTime
                                        label: نوع استخدام
            responses:
                '200':
                    description: Option updated
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                    options:
                                        type: array
                                        items:
                                            type: string
                '400':
                    $ref: '#/components/responses/Error'
                '404':
                    $ref: '#/components/responses/Error'
                '500':
                    $ref: '#/components/responses/Error'
    /options/{field}/{index}:
        delete:
            tags:
                - Options
            summary: Delete option value by index
            security:
                - bearerAuth: []
            parameters:
                - name: field
                  in: path
                  required: true
                  schema:
                      type: string
                - name: index
                  in: path
                  required: true
                  schema:
                      type: integer
            responses:
                '200':
                    description: Option deleted
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                    options:
                                        type: array
                                        items:
                                            type: string
                '400':
                    $ref: '#/components/responses/Error'
                '404':
                    $ref: '#/components/responses/Error'
                '500':
                    $ref: '#/components/responses/Error'
...
components:
    schemas:
...
        Personnel:
            type: object
            properties:
                id:
                    type: integer
                name:
                    type: string
                lname:
                    type: string
                father_name:
                    type: string
                national_id:
                    type: string
                emp_num:
                    type: string
                hire_date:
                    type: string
                emp_type:
                    type: string
                org_post:
                    type: string
                job_title:
                    type: string
                last_degree:
                    type: string
                phone:
                    type: string
                address:
                    type: string
                status:
                    type: string
                notes:
                    type: string
        PersonnelCreate:
            type: object
            required:
                - name
                - lname
                - national_id
                - emp_num
                - hire_date
                - org_post
                - job_title
                - last_degree
                - phone
                - address
            properties:
                name:
                    type: string
                    example: علی
                lname:
                    type: string
                    example: رضایی
                father_name:
                    type: string
                    example: محمد
                national_id:
                    type: string
                    example: '0123456789'
                emp_num:
                    type: string
                    example: '1001'
                hire_date:
                    type: string
                    example: '1403-01-01'
                emp_type:
                    type: string
                    example: Permanent
                org_post:
                    type: string
                    example: IT
                job_title:
                    type: string
                    example: برنامه‌نویس
                last_degree:
                    type: string
                    example: کارشناسی
                phone:
                    type: string
                    example: '09123456789'
                address:
                    type: string
                    example: تهران
                status:
                    type: string
                    example: فعال
                notes:
                    type: string
                    example: '-'
        PersonnelUpdate:
            type: object
            required:
                - name
                - lname
                - national_id
                - emp_num
                - hire_date
                - org_post
                - job_title
                - last_degree
                - phone
                - address
            properties:
                name:
                    type: string
                lname:
                    type: string
                father_name:
                    type: string
                national_id:
                    type: string
                emp_num:
                    type: string
                hire_date:
                    type: string
                emp_type:
                    type: string
                org_post:
                    type: string
                job_title:
                    type: string
                last_degree:
                    type: string
                phone:
                    type: string
                address:
                    type: string
                status:
                    type: string
                notes:
                    type: string
...
        Option:
            type: object
            properties:
                label:
                    type: string
                options:
                    type: array
                    items:
                        type: string
...
        SuccessResponse:
            type: object
            properties:
                success:
                    type: boolean
        ErrorResponse:
            type: object
            properties:
                error:
                    type: string
    responses:
        Error:
            description: Error response
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    examples:
                        badRequest:
                            value:
                                error: 'Invalid request'
                        unauthorized:
                            value:
                                error: 'دسترسی غیرمجاز! لطفاً ابتدا وارد شوید.'
                        forbidden:
                            value:
                                error: 'دسترسی غیرمجاز'
                        notFound:
                            value:
                                error: 'یافت نشد'
                        serverError:
                            value:
                                error: 'خطای سرور'